foodtales/requirements.txt
django==4.1.5
drf-yasg==1.21.4
Pillow==9.4.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==22.1
django-ckeditor==6.5.1
Markdown==3.4.1